requirement.txt
nltk>=3.2.1
newspaper3k>=0.2.6
sklearn>=0.23.0
numpy
